Rust: Refactor Tags To Lanes
