+++ rust/osm2lanes/src/transform/tags_to_lanes/bus.rs
@@ -1,0 +1,186 @@
+use super::*;
+
+pub fn bus(
+    tags: &Tags,
+    locale: &Locale,
+    oneway: bool,
+    forward_side: &mut Vec<Lane>,
+    backward_side: &mut Vec<Lane>,
+) -> ModeResult {
+    // https://wiki.openstreetmap.org/wiki/Bus_lanes
+    // 3 schemes, for simplicity we only allow one at a time
+    let tag_tree = tags.tree();
+
+    match (
+        tag_tree.get("busway").is_some(),
+        tag_tree
+            .get("lanes:psv")
+            .or_else(|| tag_tree.get("lanes:bus"))
+            .is_some(),
+        tag_tree
+            .get("bus:lanes")
+            .or_else(|| tag_tree.get("psv:lanes"))
+            .is_some(),
+    ) {
+        (false, false, false) => {}
+        (true, _, false) => bus_busway(tags, locale, oneway, forward_side, backward_side)?,
+        (false, true, false) => bus_bus_lanes(tags, locale, oneway, forward_side, backward_side)?,
+        (false, false, true) => bus_lanes_bus(tags, locale, oneway, forward_side, backward_side)?,
+        _ => {
+            return Err(RoadMsg::Unsupported {
+                description: Some("more than one bus lanes scheme used".to_owned()),
+                tags: None,
+            }
+            .into())
+        }
+    }
+
+    Ok(())
+}
+
+fn bus_busway(
+    tags: &Tags,
+    locale: &Locale,
+    _oneway: bool,
+    forward_side: &mut Vec<Lane>,
+    backward_side: &mut Vec<Lane>,
+) -> ModeResult {
+    const BUSWAY: TagKey = TagKey::from("busway");
+    if tags.is(BUSWAY, "lane") {
+        forward_side
+            .last_mut()
+            .ok_or_else(|| RoadError::unsupported_str("no forward lanes for busway"))?
+            .set_bus()?;
+        if !tags.is("oneway", "yes") && !tags.is("oneway:bus", "yes") {
+            backward_side
+                .last_mut()
+                .ok_or_else(|| RoadError::unsupported_str("no backward lanes for busway"))?
+                .set_bus()?;
+        }
+    }
+    if tags.is(BUSWAY, "opposite_lane") {
+        backward_side
+            .last_mut()
+            .ok_or_else(|| RoadError::unsupported_str("no backward lanes for busway"))?
+            .set_bus()?;
+    }
+    if tags.is(BUSWAY + "both", "lane") {
+        forward_side
+            .last_mut()
+            .ok_or_else(|| RoadError::unsupported_str("no forward lanes for busway"))?
+            .set_bus()?;
+        backward_side
+            .last_mut()
+            .ok_or_else(|| RoadError::unsupported_str("no backward lanes for busway"))?
+            .set_bus()?;
+        if tags.is("oneway", "yes") || tags.is("oneway:bus", "yes") {
+            return Err(RoadError::ambiguous_str(
+                "busway:both=lane for oneway roads",
+            ));
+        }
+    }
+    if tags.is(BUSWAY + locale.driving_side.tag(), "lane") {
+        forward_side
+            .last_mut()
+            .ok_or_else(|| RoadError::unsupported_str("no forward lanes for busway"))?
+            .set_bus()?;
+    }
+    if tags.is(BUSWAY + locale.driving_side.opposite().tag(), "lane") {
+        if tags.is("oneway", "yes") || tags.is("oneway:bus", "yes") {
+            forward_side
+                .first_mut()
+                .ok_or_else(|| RoadError::unsupported_str("no forward lanes for busway"))?
+                .set_bus()?;
+        } else {
+            return Err(RoadError::ambiguous_str(
+                "busway:BACKWARD=lane for bidirectional roads",
+            ));
+        }
+    }
+    Ok(())
+}
+
+fn bus_lanes_bus(
+    _tags: &Tags,
+    _locale: &Locale,
+    _oneway: bool,
+    _forward_side: &mut Vec<Lane>,
+    _backward_side: &mut Vec<Lane>,
+) -> ModeResult {
+    Ok(())
+}
+
+fn bus_bus_lanes(
+    tags: &Tags,
+    _locale: &Locale,
+    oneway: bool,
+    forward_side: &mut Vec<Lane>,
+    backward_side: &mut Vec<Lane>,
+) -> ModeResult {
+    let fwd_bus_spec = if let Some(s) = tags.get("bus:lanes:forward") {
+        s
+    } else if let Some(s) = tags.get("psv:lanes:forward") {
+        s
+    } else if oneway {
+        if let Some(s) = tags.get("bus:lanes") {
+            s
+        } else if let Some(s) = tags.get("psv:lanes") {
+            s
+        } else {
+            ""
+        }
+    } else {
+        ""
+    };
+    if !fwd_bus_spec.is_empty() {
+        let parts: Vec<&str> = fwd_bus_spec.split('|').collect();
+        let offset = if let Lane::Travel {
+            direction: Some(LaneDirection::Both),
+            ..
+        } = forward_side[0]
+        {
+            1
+        } else {
+            0
+        };
+        if parts.len() == forward_side.len() - offset {
+            for (idx, part) in parts.into_iter().enumerate() {
+                if part == "designated" {
+                    let direction =
+                        if let Lane::Travel { direction, .. } = forward_side[idx + offset] {
+                            direction
+                        } else {
+                            unreachable!()
+                        };
+                    forward_side[idx + offset] = Lane::Travel {
+                        direction,
+                        designated: LaneDesignated::Bus,
+                    };
+                }
+            }
+        }
+    }
+    if let Some(spec) = tags
+        .get("bus:lanes:backward")
+        .or_else(|| tags.get("psv:lanes:backward"))
+    {
+        let parts: Vec<&str> = spec.split('|').collect();
+        if parts.len() == backward_side.len() {
+            for (idx, part) in parts.into_iter().enumerate() {
+                if part == "designated" {
+                    let direction = if let Lane::Travel { direction, .. } = forward_side[idx] {
+                        direction
+                    } else {
+                        unreachable!()
+                    };
+                    backward_side[idx] = Lane::Travel {
+                        direction,
+                        designated: LaneDesignated::Bus,
+                    };
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
